refactor: siebarview
