+++ frontend/src/routes/_authenticated/applications/$id/settings/general.tsx
@@ -1,16 +1,8 @@
 import { type Application } from "@/lib/applications";
 import { useFetch } from "@/lib/api";
 import { m } from "@/lib/paraglide/messages";
-import { createFileRoute, Link } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import { ApplicationForm } from "@/components/application-form";
-import {
-	Breadcrumb,
-	BreadcrumbItem,
-	BreadcrumbLink,
-	BreadcrumbList,
-	BreadcrumbPage,
-	BreadcrumbSeparator,
-} from "@/components/ui/breadcrumb";
 
 export const Route = createFileRoute("/_authenticated/applications/$id/settings/general")({
 	component: EditApplicationPage,
@@ -28,33 +20,7 @@ function EditApplicationPage() {
 	const { data: application } = useFetch<Application>(`/applications/${id}`);
 
 	return (
-		<div className="space-y-6">
-			<Breadcrumb>
-				<BreadcrumbList>
-					<BreadcrumbItem>
-						<BreadcrumbLink
-							render={
-								<Link to="/applications">{m.pageApplications()}</Link>
-							}
-						></BreadcrumbLink>
-					</BreadcrumbItem>
-					<BreadcrumbSeparator />
-					<BreadcrumbItem>
-						<BreadcrumbLink
-							render={
-								<Link to="/applications/$id" params={{ id }}>
-									{application?.name}
-								</Link>
-							}
-						></BreadcrumbLink>
-					</BreadcrumbItem>
-					<BreadcrumbSeparator />
-					<BreadcrumbItem>
-						<BreadcrumbPage>{m.settings()}</BreadcrumbPage>
-					</BreadcrumbItem>
-				</BreadcrumbList>
-			</Breadcrumb>
-
+		<div className="p-6 space-y-6">
 			{application ? <ApplicationForm application={application} /> : <div>{m.loadingDots()}</div>}
 		</div>
 	);
